% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvialsurvival.h.R
\name{alluvialSurvival}
\alias{alluvialSurvival}
\title{Treatment Pathway Alluvial Plot}
\usage{
alluvialSurvival(timeVar)
}
\arguments{
\item{timeVar}{The name of the time variable in the dataset}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$mydataview} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Treatment Pathway Alluvial Plot
}
