% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_cancer_data}
\alias{breast_cancer_data}
\title{Breast cancer stage migration data}
\format{
A data frame
}
\usage{
breast_cancer_data
}
\description{
Stage migration test case for breast cancer.
}
\keyword{datasets}
