% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_colorectal.R
\docType{data}
\name{colorectal}
\alias{colorectal}
\title{colorectal: example dataset for the jsurvival module}
\format{
A data frame with 700 rows and 8 variables:
\describe{
\item{patient_id}{integer variable}
\item{age}{numeric variable}
\item{sex}{factor variable}
\item{old_stage}{ordered/factor variable}
\item{new_stage}{ordered/factor variable}
\item{survival_time}{numeric variable}
\item{event}{integer variable}
\item{cancer_type}{factor variable}
}
}
\usage{
colorectal
}
\description{
Example dataset distributed with the jsurvival jamovi module for demonstration and testing.
}
\keyword{datasets}
\keyword{internal}
