% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorectal_cancer_data}
\alias{colorectal_cancer_data}
\title{Colorectal cancer stage migration data}
\format{
A data frame
}
\usage{
colorectal_cancer_data
}
\description{
Stage migration test case for colorectal cancer.
}
\keyword{datasets}
