% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_data}
\alias{combined_data}
\title{Combined stage migration data}
\format{
A data frame
}
\usage{
combined_data
}
\description{
Combined stage migration test case.
}
\keyword{datasets}
