% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_data_longitudinal.R
\docType{data}
\name{data_longitudinal}
\alias{data_longitudinal}
\title{data_longitudinal: example dataset for the jsurvival module}
\format{
A data frame with 140 rows and 3 variables:
\describe{
\item{PatientID}{character variable}
\item{Time}{integer variable}
\item{Measurement}{numeric variable}
}
}
\usage{
data_longitudinal
}
\description{
Example dataset distributed with the jsurvival jamovi module for demonstration and testing.
}
\keyword{datasets}
\keyword{internal}
