% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histopathology}
\alias{histopathology}
\title{Histopathology dataset}
\format{
A data frame
}
\usage{
histopathology
}
\description{
Example pathology data with survival information for testing and demonstration.
}
\keyword{datasets}
