% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_large_performance_data.R
\docType{data}
\name{large_performance_data}
\alias{large_performance_data}
\title{large_performance_data: example dataset for the jsurvival module}
\format{
A data frame with 2000 rows and 9 variables:
\describe{
\item{patient_id}{integer variable}
\item{age}{numeric variable}
\item{sex}{factor variable}
\item{cancer_type}{factor variable}
\item{treatment_center}{factor variable}
\item{old_stage}{factor variable}
\item{new_stage}{factor variable}
\item{survival_time}{numeric variable}
\item{event}{integer variable}
}
}
\usage{
large_performance_data
}
\description{
Example dataset distributed with the jsurvival jamovi module for demonstration and testing.
}
\keyword{datasets}
\keyword{internal}
