% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_cancer_data}
\alias{lung_cancer_data}
\title{Lung cancer stage migration data}
\format{
A data frame
}
\usage{
lung_cancer_data
}
\description{
Stage migration test case for lung cancer.
}
\keyword{datasets}
