% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_melanoma.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{melanoma: example dataset for the jsurvival module}
\format{
A data frame with 205 rows and 7 variables:
\describe{
\item{time}{numeric variable}
\item{status}{numeric variable}
\item{sex}{numeric variable}
\item{age}{numeric variable}
\item{year}{numeric variable}
\item{thickness}{numeric variable}
\item{ulcer}{numeric variable}
}
}
\usage{
melanoma
}
\description{
Example dataset distributed with the jsurvival jamovi module for demonstration and testing.
}
\keyword{datasets}
\keyword{internal}
