% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stagemigration_large_performance.R
\docType{data}
\name{stagemigration_large_performance}
\alias{stagemigration_large_performance}
\title{stagemigration_large_performance: example dataset for the jsurvival module}
\format{
A data frame with 2000 rows and 9 variables:
\describe{
\item{patient_id}{integer variable}
\item{age}{numeric variable}
\item{sex}{factor variable}
\item{cancer_type}{factor variable}
\item{treatment_center}{factor variable}
\item{old_stage}{factor variable}
\item{new_stage}{factor variable}
\item{survival_time}{numeric variable}
\item{event}{integer variable}
}
}
\usage{
stagemigration_large_performance
}
\description{
Example dataset distributed with the jsurvival jamovi module for demonstration and testing.
}
\keyword{datasets}
\keyword{internal}
