% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_stats.R
\docType{data}
\name{summary_stats}
\alias{summary_stats}
\title{summary_stats: example dataset for the jsurvival module}
\format{
An object of class list of length 6.
}
\usage{
summary_stats
}
\description{
Example dataset distributed with the jsurvival jamovi module for demonstration and testing.
}
\keyword{datasets}
\keyword{internal}
