% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.h.R
\name{survival}
\alias{survival}
\title{Survival Analysis}
\usage{
survival(
  data,
  overalltime,
  tint = FALSE,
  dxdate,
  fudate,
  explanatory,
  outcome,
  outcomeLevel,
  dod,
  dooc,
  awd,
  awod,
  analysistype = "overall",
  cutp = "12, 36, 60",
  pw = FALSE,
  sc = FALSE,
  ce = FALSE,
  ch = FALSE,
  timetype,
  endplot
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{overalltime}{.}

\item{tint}{.}

\item{dxdate}{.}

\item{fudate}{.}

\item{explanatory}{.}

\item{outcome}{.}

\item{outcomeLevel}{.}

\item{dod}{.}

\item{dooc}{.}

\item{awd}{.}

\item{awod}{.}

\item{analysistype}{.}

\item{cutp}{.}

\item{pw}{.}

\item{sc}{.}

\item{ce}{.}

\item{ch}{.}

\item{timetype}{.}

\item{endplot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$todo} \tab \tab \tab \tab \tab a html \cr
\code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$medianTable} \tab \tab \tab \tab \tab a table \cr
\code{results$text5} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$uniTable} \tab \tab \tab \tab \tab a table \cr
\code{results$text7} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$survTable} \tab \tab \tab \tab \tab a table \cr
\code{results$text9} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$text8} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$plot} \tab \tab \tab \tab \tab an image \cr
\code{results$plot2} \tab \tab \tab \tab \tab an image \cr
\code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$medianTable$asDF}

\code{as.data.frame(results$medianTable)}
}
\description{
Function for Generating Summaries for Survival Analysis.
}
\examples{
\dontrun{
# example will be added
}
}
