% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeinterval.h.R
\name{timeinterval}
\alias{timeinterval}
\title{Time Interval}
\usage{
timeinterval(data, time, event, expl)
}
\arguments{
\item{data}{.}

\item{time}{.}

\item{event}{.}

\item{expl}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$text1html} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Time Interval
}
