% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_longitudinal}
\alias{data_longitudinal}
\title{Longitudinal data}
\format{
A data frame
}
\usage{
data_longitudinal
}
\description{
Example longitudinal data for testing and demonstration.
}
\keyword{datasets}
