% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{large_performance_data}
\alias{large_performance_data}
\title{Large performance stage migration data}
\format{
A data frame
}
\usage{
large_performance_data
}
\description{
Large performance stage migration test case.
}
\keyword{datasets}
