% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma survival dataset}
\format{
A data frame
}
\usage{
melanoma
}
\description{
Melanoma survival dataset for testing and demonstration.
}
\keyword{datasets}
