% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisurvival.h.R
\name{multisurvival}
\alias{multisurvival}
\title{Multivariable Survival Analysis}
\usage{
multisurvival(
  data,
  elapsedtime,
  tint = FALSE,
  dxdate,
  fudate,
  outcome,
  outcomeLevel,
  dod,
  dooc,
  awd,
  awod,
  explanatory,
  contexpl,
  multievent = FALSE,
  analysistype = "overall",
  timetypedata = "ymd",
  timetypeoutput = "months",
  uselandmark = FALSE,
  landmark = 3,
  hr = FALSE,
  sty = "t1",
  ph_cox = FALSE,
  calculateRiskScore = FALSE,
  plotRiskGroups = FALSE,
  ac = FALSE,
  adjexplanatory,
  km = FALSE,
  endplot = 60,
  byplot = 12,
  ci95 = FALSE,
  risktable = FALSE,
  censored = FALSE,
  pplot = TRUE,
  ac_method = "average",
  ac_summary = FALSE,
  ac_timepoints = "12, 36, 60",
  ac_compare = FALSE,
  reduced_explanatory,
  compare_models = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{elapsedtime}{.}

\item{tint}{.}

\item{dxdate}{.}

\item{fudate}{.}

\item{outcome}{.}

\item{outcomeLevel}{.}

\item{dod}{.}

\item{dooc}{.}

\item{awd}{.}

\item{awod}{.}

\item{explanatory}{.}

\item{contexpl}{.}

\item{multievent}{.}

\item{analysistype}{.}

\item{timetypedata}{select the time type in data}

\item{timetypeoutput}{select the time type in output}

\item{uselandmark}{.}

\item{landmark}{.}

\item{hr}{.}

\item{sty}{.}

\item{ph_cox}{.}

\item{calculateRiskScore}{.}

\item{plotRiskGroups}{.}

\item{ac}{.}

\item{adjexplanatory}{.}

\item{km}{.}

\item{endplot}{.}

\item{byplot}{.}

\item{ci95}{.}

\item{risktable}{.}

\item{censored}{.}

\item{pplot}{.}

\item{ac_method}{Method for computing adjusted survival curves}

\item{ac_summary}{Calculate and display summary statistics for adjusted
curves}

\item{ac_timepoints}{Timepoints for calculating summary statistics
(comma-separated)}

\item{ac_compare}{Perform statistical comparison between adjusted curves}

\item{reduced_explanatory}{.}

\item{compare_models}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$todo} \tab \tab \tab \tab \tab a html \cr
\code{results$text} \tab \tab \tab \tab \tab a html \cr
\code{results$text2} \tab \tab \tab \tab \tab a html \cr
\code{results$plot} \tab \tab \tab \tab \tab an image \cr
\code{results$plot3} \tab \tab \tab \tab \tab an image \cr
\code{results$cox_ph} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$plot8} \tab \tab \tab \tab \tab an image \cr
\code{results$mydataview} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$riskSummaryTable} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$riskScoreText} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$plotKM} \tab \tab \tab \tab \tab an image \cr
\code{results$riskScoreTable} \tab \tab \tab \tab \tab a table \cr
\code{results$riskScoreMetrics} \tab \tab \tab \tab \tab a html \cr
\code{results$riskGroupPlot} \tab \tab \tab \tab \tab an image \cr
\code{results$calculatedtime} \tab \tab \tab \tab \tab an output \cr
\code{results$outcomeredefined} \tab \tab \tab \tab \tab an output \cr
\code{results$addRiskScore} \tab \tab \tab \tab \tab an output \cr
\code{results$addRiskGroup} \tab \tab \tab \tab \tab an output \cr
\code{results$mydataview_plot_adj} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$plot_adj} \tab \tab \tab \tab \tab an image \cr
\code{results$mydataview_calculateAdjustedStats} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$adjustedSummaryTable} \tab \tab \tab \tab \tab a table \cr
\code{results$adjustedComparison} \tab \tab \tab \tab \tab a html \cr
\code{results$model_comparison} \tab \tab \tab \tab \tab a html \cr
\code{results$reduced_model_metrics} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$riskScoreTable$asDF}

\code{as.data.frame(results$riskScoreTable)}
}
\description{
Function for Multivariable Survival Analysis using Cox-regression.
}
\examples{
\donttest{
# example will be added
}
}
