% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisurvival.h.R
\name{multisurvival}
\alias{multisurvival}
\title{Multivariate Survival Analysis}
\usage{
multisurvival(
  data,
  elapsedtime,
  tint = FALSE,
  dxdate,
  fudate,
  explanatory,
  outcome,
  outcomeLevel,
  dod,
  dooc,
  awd,
  awod,
  multievent = FALSE,
  analysistype = "overall",
  timetypedata = "ymd",
  timetypeoutput = "months",
  hr = FALSE,
  sty = "t1",
  ac = FALSE,
  adjexplanatory
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{elapsedtime}{.}

\item{tint}{.}

\item{dxdate}{.}

\item{fudate}{.}

\item{explanatory}{.}

\item{outcome}{.}

\item{outcomeLevel}{.}

\item{dod}{.}

\item{dooc}{.}

\item{awd}{.}

\item{awod}{.}

\item{multievent}{.}

\item{analysistype}{.}

\item{timetypedata}{select the time type in data}

\item{timetypeoutput}{select the time type in output}

\item{hr}{.}

\item{sty}{.}

\item{ac}{.}

\item{adjexplanatory}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$todo} \tab \tab \tab \tab \tab a html \cr
\code{results$text} \tab \tab \tab \tab \tab a html \cr
\code{results$text2} \tab \tab \tab \tab \tab a html \cr
\code{results$plot} \tab \tab \tab \tab \tab an image \cr
\code{results$plot3} \tab \tab \tab \tab \tab an image \cr
\code{results$plot4} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Function for Multivariate Survival Analysis using Cox-regression.
}
\examples{
\dontrun{
# example will be added
}
}
