% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problematic_data}
\alias{problematic_data}
\title{Problematic stage migration data}
\format{
A data frame
}
\usage{
problematic_data
}
\description{
Problematic stage migration test case with edge cases.
}
\keyword{datasets}
