% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_sample_data}
\alias{small_sample_data}
\title{Small sample stage migration data}
\format{
A data frame
}
\usage{
small_sample_data
}
\description{
Small sample stage migration test case.
}
\keyword{datasets}
