% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_stats}
\alias{summary_stats}
\title{Summary statistics stage migration data}
\format{
A data frame
}
\usage{
summary_stats
}
\description{
Summary statistics for stage migration analysis.
}
\keyword{datasets}
